Allow voters to change name once in a room

+++ lambda/disconnect/index.js
@@ -59,7 +59,7 @@ exports.handler = async (event) => {
                 return voterData;
         });
 
-        if (votersData.length === 0) {
+        if (roomID !== "noroom" && votersData.length === 0) {
             promises.push(
                 ddb
                     .deleteItem({
